refactor retail-entry;add agent base
