--- source/lambda/online/functions/lambda_retail_tools/lambda_customer_complain/customer_complain.py
@@ -1,5 +1,0 @@
-# customer complain
-
-def lambda_handler(event_body,context=None):
-    raise NotImplementedError
-    # return {"code":0, "result":event_body['kwargs']['response']}
