--- source/lambda/online/functions/lambda_retail_tools/lambda_daily_reception/daily_reception.py
@@ -1,4 +1,0 @@
-# daily reception tool
-
-def lambda_handler(event_body,context=None):
-    return {"code":0, "result":event_body['kwargs']['response']}
