--- source/lambda/online/functions/lambda_retail_tools/lambda_goods_exchange/goods_exchange.py
@@ -1,5 +1,0 @@
-# goods exchange
-
-def lambda_handler(event_body,context=None):
-    raise NotImplementedError
-    return {"code":0, "result":event_body['kwargs']['response']}
