--- source/lambda/online/functions/lambda_retail_tools/lambda_human/human.py
@@ -1,11 +1,0 @@
-# test tool
-
-
-def size_tool():
-    return "L"
-
-def lambda_handler(event_body,context=None):
-    # retriever
-    # llm_generate
-    # size tool
-    return {"code":0, "result":"I am a retail bot.", "name": "retail_bot"}
