--- source/lambda/online/functions/lambda_retail_tools/lambda_order_info/retail/intent_data/intent_ingestion_data.jsonl
@@ -1,114 +1,0 @@
-{"question": "亲，这个大人能穿吗\n165身高的话可以换165m吗", "answer": {"intent": "尺码查询"}}
-{"question": "230拍哪个尺码", "answer": {"intent": "尺码查询"}}
-{"question": "平常41吗，这款鞋需要多少码", "answer": {"intent": "尺码查询"}}
-{"question": "平时穿40，码数偏大么", "answer": {"intent": "尺码查询"}}
-{"question": "亲，这个大人能穿吗\n165身高的话可以换165m吗", "answer": {"intent": "尺码查询"}}
-{"question": "尺码是怎样的", "answer": {"intent": "尺码查询"}}
-{"question": "这个事情侣款吗\n情侣款的话，女款，发我一下", "answer": {"intent": "尺码查询"}}
-{"question": "平常41吗，这款鞋需要多少码", "answer": {"intent": "尺码查询"}}
-{"question": "177 65kg多大", "answer": {"intent": "尺码查询"}}
-{"question": "平时穿40，码数偏大么", "answer": {"intent": "尺码查询"}}
-{"question": "155.厘米125斤", "answer": {"intent": "尺码查询"}}
-{"question": "亲，这个大人能穿吗\n165身高的话可以换165m吗", "answer": {"intent": "尺码查询"}}
-{"question": "158 54要什么号呀", "answer": {"intent": "尺码查询"}}
-{"question": "脚瘦的宝宝\n适合不", "answer": {"intent": "尺码查询"}}
-{"question": "我们脚长20.6", "answer": {"intent": "尺码查询"}}
-{"question": "平常穿41拍42会不会太大", "answer": {"intent": "尺码查询"}}
-{"question": "鞋码偏大还是偏小", "answer": {"intent": "尺码查询"}}
-{"question": "已经最小码了", "answer": {"intent": "尺码查询"}}
-{"question": "38码不是240吗？怎会235？？", "answer": {"intent": "尺码查询"}}
-{"question": "线上买的不可以积分吗", "answer": {"intent": "促销查询"}}
-{"question": "这个券我抢到了，下单的时候自动使用吗", "answer": {"intent": "促销查询"}}
-{"question": "这个160元优惠券使用有限制吗", "answer": {"intent": "促销查询"}}
-{"question": "线上买的不可以积分吗", "answer": {"intent": "促销查询"}}
-{"question": "评论有惊喜吗", "answer": {"intent": "促销查询"}}
-{"question": "https://img.alicdn.com/imgextra/i4/O1CN01fTnEMC1CKRUnHH54M_!!0-amp.jpg", "answer": {"intent": "规则回复"}}
-{"question": "http://item.taobao.com/item.htm?id=751505986415", "answer": {"intent": "规则回复"}}
-{"question": "3405169803836764400", "answer": {"intent": "规则回复"}}
-{"question": "还是避重就轻，看来这是你们的习惯\n买你们的东西可以相当于一次考试了", "answer": {"intent": "客户抱怨"}}
-{"question": "还是避重就轻，看来这是你们的习惯\n买你们的东西可以相当于一次考试了", "answer": {"intent": "客户抱怨"}}
-{"question": "给催一下啊", "answer": {"intent": "客户抱怨"}}
-{"question": "很浪费时间 出库的时候也不看清楚", "answer": {"intent": "客户抱怨"}}
-{"question": "谢谢", "answer": {"intent": "日常接待"}}
-{"question": "好的", "answer": {"intent": "日常接待"}}
-{"question": "在吗", "answer": {"intent": "日常接待"}}
-{"question": "好吧", "answer": {"intent": "日常接待"}}
-{"question": "呵呵", "answer": {"intent": "日常接待"}}
-{"question": "你好", "answer": {"intent": "日常接待"}}
-{"question": "好的，谢谢", "answer": {"intent": "日常接待"}}
-{"question": "还是原来那款感觉好点", "answer": {"intent": "日常接待"}}
-{"question": "我找一下哈", "answer": {"intent": "日常接待"}}
-{"question": "怎么操作呢", "answer": {"intent": "日常接待"}}
-{"question": "哦好的，谢谢", "answer": {"intent": "日常接待"}}
-{"question": "王晓燕15145338269", "answer": {"intent": "日常接待"}}
-{"question": "在吗", "answer": {"intent": "日常接待"}}
-{"question": "正确", "answer": {"intent": "日常接待"}}
-{"question": "在不", "answer": {"intent": "日常接待"}}
-{"question": "我也是卡点", "answer": {"intent": "日常接待"}}
-{"question": "杨幂同款裤子有吗", "answer": {"intent": "商品推荐"}}
-{"question": "可以定制工作服吗", "answer": {"intent": "商品推荐"}}
-{"question": "能给我推荐一套衣服吗", "answer": {"intent": "商品推荐"}}
-{"question": "有别人家图片里的东西吗？", "answer": {"intent": "商品推荐"}}
-{"question": "哦 我喜欢厚底的鞋  一般厚度就行", "answer": {"intent": "商品推荐"}}
-{"question": "这款鞋没有增高吧 不喜欢增高的", "answer": {"intent": "商品信息查询"}}
-{"question": "这款呢 鞋跟大约多少", "answer": {"intent": "商品信息查询"}}
-{"question": "这款鞋底大概多厚", "answer": {"intent": "商品信息查询"}}
-{"question": "请问这个怎么购买呀\n显示有货，但是无法购买", "answer": {"intent": "商品信息查询"}}
-{"question": "这体恤夏天穿热不热", "answer": {"intent": "商品信息查询"}}
-{"question": "这个鞋哪个连接", "answer": {"intent": "商品信息查询"}}
-{"question": "裤子有连接吗", "answer": {"intent": "商品信息查询"}}
-{"question": "这个是没有货了吗", "answer": {"intent": "商品信息查询"}}
-{"question": "黑色的175什么时候会有？", "answer": {"intent": "商品信息查询"}}
-{"question": "差不多款式的有吗？", "answer": {"intent": "商品信息查询"}}
-{"question": "你家的鞋适合脚瘦还是脚胖的", "answer": {"intent": "商品信息查询"}}
-{"question": "我一个孩子瘦，一个孩子胖", "answer": {"intent": "商品信息查询"}}
-{"question": "脚特别胖的，脚背宽高的呢", "answer": {"intent": "商品信息查询"}}
-{"question": "这是去年的款\n今年的新款呢\n啥时候上来", "answer": {"intent": "商品信息查询"}}
-{"question": "发黄的厉害", "answer": {"intent": "商品信息查询"}}
-{"question": "三包期是多久\n", "answer": {"intent": "商品信息查询"}}
-{"question": "有啥办法去黄嘛\n我没洗过呢", "answer": {"intent": "商品信息查询"}}
-{"question": "咋去黄\n就这么简单a", "answer": {"intent": "商品信息查询"}}
-{"question": "这款鞋啥材质", "answer": {"intent": "商品信息查询"}}
-{"question": "链接给我", "answer": {"intent": "商品信息查询"}}
-{"question": "这款鞋没有增高吧 不喜欢增高的", "answer": {"intent": "商品信息查询"}}
-{"question": "这款呢 鞋跟大约多少？", "answer": {"intent": "商品信息查询"}}
-{"question": "这个纯黑的是三公分的吗", "answer": {"intent": "商品信息查询"}}
-{"question": "这个属于运动鞋吧 不是休息鞋吧  可以这样参考吧", "answer": {"intent": "商品信息查询"}}
-{"question": "不不是让我参考的吗\n我要分便厚度", "answer": {"intent": "商品信息查询"}}
-{"question": "哦这个款我可以分便成2-3 公分吧", "answer": {"intent": "商品信息查询"}}
-{"question": "过完年鞋子还没打折吗", "answer": {"intent": "商品信息查询"}}
-{"question": "这是什么款", "answer": {"intent": "商品信息查询"}}
-{"question": "鞋跟多少cm", "answer": {"intent": "商品信息查询"}}
-{"question": "鞋子里面都没有塞东西的吗", "answer": {"intent": "商品信息查询"}}
-{"question": "夏款是那个", "answer": {"intent": "商品信息查询"}}
-{"question": "可以吧黑的换成燕麦色吗", "answer": {"intent": "商品信息查询"}}
-{"question": "你家鞋子开胶了\n怎么处理", "answer": {"intent": "商品质量问题"}}
-{"question": "如果是在你家买的鞋子，出现质量问题你们怎么处理", "answer": {"intent": "商品质量问题"}}
-{"question": "你好 我双十一买的这个鞋 过年才穿 洗了两次 就这样了", "answer": {"intent": "商品质量问题"}}
-{"question": "我上次买的鞋上有污渍脏了你们之前的客服为我办理了退货申请我也寄回去了，但我还是想要这双鞋现在重新拍一双，这次给我检查好不要在有这种情况了\n客服刚才说这次帮我备注一下一定检查好了", "answer": {"intent": "商品质量问题"}}
-{"question": "没洗过！现在黄的一塌糊涂", "answer": {"intent": "商品质量问题"}}
-{"question": "退吧 我不买这个鞋子了\n你鞋子有问题 我承担运费？", "answer": {"intent": "退换货"}}
-{"question": "麻烦同意下我的换货申请", "answer": {"intent": "退换货"}}
-{"question": "请问你们是哪里发货", "answer": {"intent": "物流规则"}}
-{"question": "我拼多多买的为什么是你们那里发货？", "answer": {"intent": "物流规则"}}
-{"question": "能发顺丰嘛？", "answer": {"intent": "物流规则"}}
-{"question": "你好！我的换货已经收到，垫付了的运费可以返了吧", "answer": {"intent": "物流规则"}}
-{"question": "点完确认收货后，没有仅退款这项啊", "answer": {"intent": "物流规则"}}
-{"question": "可以顺丰特快吗", "answer": {"intent": "物流规则"}}
-{"question": "你帮我查一下这个是你们发出的吗？", "answer": {"intent": "物流信息查询"}}
-{"question": "帮我查一下这个物流单号", "answer": {"intent": "物流信息查询"}}
-{"question": "可以顺丰特快吗", "answer": {"intent": "物流信息查询"}}
-{"question": "急用，有没有次日达物流，山东的", "answer": {"intent": "物流信息查询"}}
-{"question": "这个是你们发出的吗", "answer": {"intent": "物流信息查询"}}
-{"question": "能查到吗\nSF1664371953288", "answer": {"intent": "物流信息查询"}}
-{"question": "你帮我查一下这个是你们发出的吗", "answer": {"intent": "物流信息查询"}}
-{"question": "帮我查一下这个物流单号", "answer": {"intent": "物流信息查询"}}
-{"question": "给我加急一下吧，这马上过生日了，要不我就线下买了", "answer": {"intent": "物流信息查询"}}
-{"question": "你好麻烦帮我看一下这个售后", "answer": {"intent": "物流信息查询"}}
-{"question": "我已下单，请问何时发货？", "answer": {"intent": "物流信息查询"}}
-{"question": "这个忘机用券了, 我重拍吧", "answer": {"intent": "下单流程"}}
-{"question": "券 一天放多少呢？", "answer": {"intent": "下单流程"}}
-{"question": "订单，可以改一下吗？", "answer": {"intent": "下单流程"}}
-{"question": "麻烦看下前期的聊天记录哈", "answer": {"intent": "信息缺失"}}
-{"question": "人工客服", "answer": {"intent": "转人工"}}
-{"question": "人工", "answer": {"intent": "转人工"}}
