--- source/lambda/online/functions/lambda_retail_tools/lambda_order_info/retail/size/偏小女装-商品ID.txt
@@ -1,208 +1,0 @@
-﻿718506064076
-718301051246
-702341369744
-702383009108
-708354787497
-714132763664
-711107896116
-717962808767
-705050740251
-648440458326
-653386076482
-653658757607
-654462770214
-655063904208
-657438021797
-659297224498
-659558130443
-660097922484
-660449167557
-660691444866
-661453664935
-661481970924
-662372396194
-662381688042
-662621031859
-662791153355
-662791645881
-663193804862
-663275400916
-663718173749
-664057970557
-664070178458
-664273768904
-664418539648
-665016868985
-665045838855
-665416961408
-665767966287
-666197882944
-666565947268
-666766079560
-666768375975
-667839400330
-668631462773
-668858702606
-668865834006
-668866258902
-670452638186
-672402147382
-672613897651
-673690238083
-673911012186
-674270544743
-674368248414
-674691894708
-674744789599
-674748237263
-674765721273
-674767113490
-675086798338
-675333944043
-675418220033
-675869203662
-675931636044
-675957232871
-676020605972
-676247522033
-677354696715
-677439002691
-677671289223
-678067833051
-678359362209
-678860354901
-679219031782
-679400357846
-679564144529
-679605988544
-680265372092
-680569785762
-680597092347
-680908462562
-680958441388
-681232883924
-681571203042
-681579895747
-681850537379
-681862354987
-681863782221
-681904046908
-682004073156
-682055764503
-682295238027
-682327109329
-682364725762
-682389531913
-682607463540
-682687442729
-682866645139
-682936971991
-682357813329
-683003279606
-683599930342
-683817504231
-684327123010
-684404736522
-684613760971
-685207522731
-685255550886
-685551391750
-685551579319
-685552511614
-685652511614
-691748326985
-691881455878
-691928964240
-692039548306
-692379204344
-692380632688
-692522026814
-692777783708
-692789211501
-693300315134
-693665388700
-694087885471
-694214319161
-694863697949
-695541740403
-699551110114
-702378572637
-702634637902
-703145371216
-703748150041
-705600319915
-706649066738
-708216272495
-708223668339
-709230789518
-709634580201
-709644307092
-718109090155
-716807768402
-714524024031
-699105071519
-644758073444
-644437277464
-718920302308
-718352933406
-715102442559
-703368317757
-719280111392
-719056413225
-716955737333
-712936547025
-712255114436
-712067489615
-718346518711
-719136374226
-702057084075
-705815672045
-701764244540
-702528603290
-719234118874
-706538135873
-705555722715
-714839303110
-702756119927
-703148799629
-718476596205
-713748154526
-725284063657
-706341138666
-708702992903
-715058654174
-702682749381
-722261835487
-719230626465
-714439153019
-698175415875
-727125364308
-711696467436
-728634672238
-728649036981
-698174455699
-706316418670
-719623860577
-724962897348
-732595455842
-731222266750
-736440468114
-739378988354
-729127015754
-736741514822
-733216064785
-729163195165
-736703274448
-736622225859
-732593919427
-736856243716
-746832813909
-736623697494
-736732582786
-733423781945
-737225547606
-734464673627
-749104080677
-733420505628
-749308912823
-733871078226
