--- source/lambda/online/functions/lambda_retail_tools/lambda_order_info/retail/size/偏小男装-商品ID.txt
@@ -1,54 +1,0 @@
-﻿653341550645
-659650109696
-663261020840
-663687945863
-664417187017
-664933718189
-665424009035
-665847185747
-666197514169
-666402582886
-666905362446
-666914758927
-667040028204
-668913443134
-668949289421
-669669961615
-671334668310
-672332817154
-673657634406
-673711630578
-674326772481
-674337112233
-674708361864
-674711849633
-674839976746
-676619897753
-681987131371
-681991132977
-682011956790
-682957187062
-684754055710
-692382259108
-685247750297
-695537240789
-710961815341
-704504342528
-718201993140
-712509169437
-714430453419
-712918847298
-718964517709
-672205028475
-674077769991
-679110151447
-679695892648
-724101353002
-719138070677
-697901006623
-715137406563
-726321724017
-694782805395
-721918957764
-736820747253
-750391088500
