--- source/lambda/online/functions/lambda_retail_tools/lambda_order_info/retail/size/情侣装-ID.txt
@@ -1,41 +1,0 @@
-﻿611536082883
-652050649811
-659140773308
-667034876909
-668133956221
-669242845363
-669279082672
-670039122074
-670390806933
-672599044216
-672661834772
-672662562109
-672825188542
-672828116415
-672982416567
-673403629253
-673689342408
-673689494024
-680276112711
-682470079121
-682474739600
-684314519299
-684916273306
-609978427207
-696768554486
-696531501465
-706555490391
-713711301149
-712303388515
-706339345375
-725463600651
-731894850334
-734865939148
-732019088798
-740055308367
-740229797487
-741833104488
-747909122614
-747572210347
-751133599235
-755655374617
