--- source/lambda/online/functions/lambda_retail_tools/lambda_order_info/retail/size/童鞋-商品ID.txt
@@ -1,460 +1,0 @@
-﻿728697400611
-732227375430
-714684892195
-725234026571
-714683708380
-695052755826
-708347133118
-699700491978
-698910492200
-699438706025
-703518599023
-699234609503
-699395663488
-703518195311
-699557187324
-698765568042
-698898222414
-699718475947
-698921220458
-718363490746
-699398243474
-699710467642
-699459138183
-699722483350
-699145846128
-703283190521
-718182305452
-708557586670
-708356045815
-699392939816
-699138622160
-699314442900
-699708535390
-703511003061
-702763348562
-718363478759
-717943168221
-718363654183
-718554443879
-646013648025
-647459795831
-649516291597
-652044553691
-656001884612
-656451193489
-656451325668
-657421814142
-657814871779
-658774103265
-658777107340
-660291416950
-660293008832
-661096274003
-661244708925
-661534720287
-661955089824
-662125531230
-662319326701
-662408383797
-665787793416
-665788297560
-665788801628
-666072578423
-666634857921
-667348027262
-667760612132
-667763800694
-668150821996
-668153537357
-668426865758
-668611956407
-668612140773
-668615244270
-668875329935
-669003077341
-669003893989
-669004421988
-669345030469
-669345535074
-669345762253
-669346986215
-669501716279
-669572347220
-669697899383
-669698491141
-669770517770
-669784392589
-669784776303
-669785180719
-669785204088
-670223542711
-670224714151
-670225086586
-670226458471
-670227022914
-670578727818
-671008892958
-671081700208
-671384125905
-671386869979
-671467493573
-672069687683
-672139933083
-672314572170
-672578092389
-672578636953
-673120972862
-673470441511
-673472513414
-673597145184
-673626287920
-673791120433
-673791612108
-673800230199
-674139087876
-674162409996
-674167757986
-674168185338
-674264483386
-674264535215
-674490654806
-674490878619
-674491226605
-674491230191
-674491598394
-674537705898
-674537941218
-674825083033
-674826151065
-674827771635
-674829627241
-674829679409
-676535596359
-676536144385
-676536276344
-676536784954
-676537460108
-676537624600
-676537664082
-676537792928
-676538184536
-676538292044
-676538300656
-676538696104
-676860177482
-676902353754
-676903101817
-676903637390
-677133558966
-677134038090
-677134354495
-677134382910
-677207458813
-677214494701
-677216602946
-677217018580
-677468127168
-677468467023
-677469279002
-677469639779
-677548807231
-677549999304
-677551503889
-677551779344
-677551823931
-677552439396
-677552443011
-678369350577
-678708191945
-679016580879
-679083980367
-679240372729
-679243316250
-679245224907
-679253760072
-679256240380
-679263832990
-679282940464
-679283440506
-679284072788
-679286844022
-679288652104
-679290036415
-679290656936
-679376865467
-679501866073
-679599209925
-679599685361
-679600249519
-679600549287
-679601297227
-679602045540
-679602673526
-679623309062
-679636221806
-679636745266
-679642293822
-679643841128
-679643909573
-679644661455
-679644829719
-679645357835
-679646161980
-679649293295
-679898178868
-679917594385
-679917802486
-679937690142
-679943566246
-679945086788
-679945446474
-679947430491
-679950514881
-679973406643
-679974354111
-679974754230
-680006755276
-680079528105
-680080476428
-680080980959
-680207454529
-680225139869
-680225707666
-680226215779
-680227295106
-680230955153
-680231319820
-680232131141
-680239751535
-680243815081
-680270923206
-680274911366
-680275515753
-680276643444
-680277119950
-680304519476
-680432741867
-680435665229
-680738218636
-680739270754
-681059915476
-681064523848
-681064807781
-681996088509
-682348197989
-682348393712
-682348437596
-682349465406
-682642126758
-682642214859
-682642234524
-682642530079
-682642538044
-682957771716
-682957895419
-682957979246
-682958007259
-683361008879
-683361220877
-683383116599
-683387160792
-683388852487
-683390172810
-683391296960
-683603169466
-683723885615
-683735057856
-683893750450
-683897374240
-684022382043
-684029366283
-684035582301
-684205727118
-684205783473
-684206595022
-684308675385
-684308835448
-684312407402
-684328539611
-684453029325
-691121518841
-691388663077
-691521948284
-691965812332
-691965972102
-691966076844
-692033608709
-692054804103
-692116470920
-692292121773
-692358893398
-692359113533
-692359965966
-692360281589
-692381305679
-692381505006
-692381901807
-692556374041
-692556758875
-692557210762
-692645662218
-692823707517
-692892159285
-692914151871
-694528549001
-694529289143
-695051675392
-695895849835
-696403327179
-699136534292
-699204309608
-699217289311
-699435610784
-699436918309
-699706923557
-699728663290
-699730931186
-702745332729
-702761140601
-703511463687
-703511999222
-703513111853
-706701182779
-708345245375
-710277070503
-703500103580
-703047657055
-699570407807
-712515997921
-712428319150
-712370716599
-712369932110
-708776867729
-708774399946
-708770027835
-708768787780
-708556054874
-708546910923
-708355125779
-708352193736
-708351157937
-708346665506
-708069868249
-708067944107
-703280326805
-703056857332
-703055649276
-703052817218
-703051429457
-703046049343
-702758536310
-702754476473
-702458773607
-699446922347
-699143974444
-698903685460
-695892433962
-692914091190
-692644678693
-692382037993
-715114678054
-718561208690
-715320507406
-713031455499
-713028375956
-712819142627
-719315903046
-718554651168
-715116274893
-712626517416
-723852077166
-692033340929
-703519115972
-714921877404
-724855208198
-692381485688
-725078917438
-725175150656
-725232110458
-725404107444
-714923205174
-732071954274
-715119094137
-714921025690
-724850028642
-735606936434
-714940637665
-728647056333
-731742744996
-729210915945
-735834924521
-736140363147
-725227090044
-695052211620
-731942129958
-736145342816
-731743484449
-728859485396
-729165559770
-731743380637
-731942197663
-729160531995
-731743532282
-737289041932
-731942201833
-731942177897
-728698928472
-735610232505
-732227743689
-729216347777
-728910033587
-728919521542
-729155439659
-725225698688
-732227919154
-728856333691
-725069393963
-738699740967
-737287033450
-737283245482
-735647324373
-731942225866
-736030893132
-725227758233
-738698608527
-736144186907
-725224858562
-736874915662
-735657896967
-738982622471
-725235830161
-735418778842
-732072198974
-728906181183
-729000626855
-737082292518
-738851649047
-728995406795
-739080879033
-737395246998
-737277213593
-731942289437
-736050067824
-738696792310
-728643236229
-737384322162
-743984665139
-743985205191
-728852613144
-738679584323
-738965090463
-729167079144
-743984229133
-737382406624
-738880813526
-744068778965
-737107948850
-748528873960
-728649320134
-748603226831
