--- source/lambda/online/functions/lambda_retail_tools/lambda_product_aftersales/product_aftersales.py
@@ -1,5 +1,0 @@
-# goods after sales
-
-def lambda_handler(event_body,context=None):
-    raise NotImplementedError
-    # return {"code":0, "result":event_body['kwargs']['response']}
