+++ source/lambda/online/functions/lambda_retail_tools/transfer.py
@@ -1,0 +1,3 @@
+# retail transfer
+def lambda_handler(event_body,context=None):
+    return {"code":0, "result":"您好,我是安踏官方客服,很高兴为您服务。请问您有什么需要帮助的吗?"}
