+++ source/lambda/online/lambda_main/main_utils/online_entries/retail_entry.py
@@ -9,24 +9,20 @@ from langgraph.graph import StateGraph,END
 from common_logic.common_utils.lambda_invoke_utils import invoke_lambda,node_monitor_wrapper
 from common_logic.common_utils.python_utils import update_nest_dict,add_messages
 from common_logic.common_utils.constant import (
-    LLMTaskType
+    LLMTaskType,
+    ToolRuningMode,
+    SceneType
 )
 
 from functions.lambda_retail_tools.lambda_product_information_search.product_information_search import goods_dict
-from functions.tool_execute_result_format import format_tool_call_results
-from functions.tool_calling_parse import parse_tool_calling as _parse_tool_calling
 
 from lambda_main.main_utils.parse_config import parse_retail_entry_config
 from common_logic.common_utils.lambda_invoke_utils import send_trace,is_running_local
-from common_logic.common_utils.exceptions import (
-    ToolNotExistError,
-    ToolParameterNotExistError,
-    MultipleToolNameError,
-    ToolNotFound
-)
 from common_logic.common_utils.logger_utils import get_logger
 from common_logic.common_utils.serialization_utils import JSONEncoder
 from common_logic.common_utils.s3_utils import download_file_from_s3,check_local_folder
+from lambda_main.main_utils.online_entries.agent_base import build_agent_graph,tool_execution
+from functions import get_tool_by_name,init_retail_tools
 
 order_info_path = "/tmp/functions/retail_tools/lambda_order_info/order_info.json"
 check_local_folder(order_info_path)
@@ -35,7 +31,8 @@ order_dict = json.load(open(order_info_path))
 
 logger = get_logger('retail_entry')
 
-goods_info_tag = "商品信息"
+init_retail_tools()
+# goods_info_tag = "商品信息"
 
 class ChatbotState(TypedDict):
     chatbot_config: dict # chatbot config
@@ -72,13 +69,14 @@ class ChatbotState(TypedDict):
     agent_recursion_limit: int # agent recursion limit
     current_agent_recursion_num: int
     enable_trace: bool
+    goods_info_tag = "商品信息"
 
 ####################
 # nodes in lambdas #
 ####################
 
 @node_monitor_wrapper
-def query_preprocess_lambda(state: ChatbotState):
+def query_preprocess(state: ChatbotState):
     output:str = invoke_lambda(
         event_body={**state,"chat_history":[]},
         lambda_name="Online_Query_Preprocess",
@@ -93,7 +91,7 @@ def query_preprocess_lambda(state: ChatbotState):
         }
 
 @node_monitor_wrapper
-def intention_detection_lambda(state: ChatbotState):
+def intention_detection(state: ChatbotState):
     intention_fewshot_examples = invoke_lambda(
         lambda_module_path='lambda_intention_detection.intention',
         lambda_name="Online_Intention_Detection",
@@ -112,7 +110,7 @@ def intention_detection_lambda(state: ChatbotState):
         }
 
 @node_monitor_wrapper
-def agent_lambda(state: ChatbotState):
+def agent(state: ChatbotState):
     goods_info = state.get('goods_info',None) or ""
     agent_chat_history = state.get('agent_chat_history',"")
     if agent_chat_history and hasattr(agent_chat_history[-1],'additional_kwargs'):
@@ -158,362 +156,347 @@ def agent_lambda(state: ChatbotState):
                 "current_agent_recursion_num": current_agent_recursion_num
             }
 
-    current_agent_output:dict = invoke_lambda(
-        event_body={
-            **state,
-            "other_chain_kwargs":{
+    # deal with once tool calling
+    if state['agent_recursion_validation'] and state['parse_tool_calling_ok'] and state['agent_chat_history']:
+        tool_execute_res = state['agent_chat_history'][-1]['additional_kwargs']['raw_tool_call_results'][0]
+        tool_name = tool_execute_res['name']
+        output = tool_execute_res['output']
+        tool = get_tool_by_name(tool_name,scene=SceneType.RETAIL)
+        if tool.running_mode == ToolRuningMode.ONCE:
+            send_trace("once tool")
+            return {
+                "answer": str(output['result']),
+                "is_current_tool_calling_once": True
+            }
+    
+    other_chain_kwargs = {
                 "goods_info": goods_info,
                 "create_time": state['create_time'],
                 "current_agent_recursion_num":state['current_agent_recursion_num']
-                }
-        },
-        lambda_name="Online_Agent",
-        lambda_module_path="lambda_agent.agent",
-        handler_name="lambda_handler"
-    )
-    current_agent_recursion_num = state['current_agent_recursion_num'] + 1
-    send_trace(f"\n\n**current_agent_output:** \n{json.dumps(current_agent_output['agent_output'],ensure_ascii=False,indent=2)}\n\n **current_agent_recursion_num:** {current_agent_recursion_num}", state["stream"], state["ws_connection_id"])
-    return {
-        "current_agent_output": current_agent_output,
-        "current_agent_recursion_num": current_agent_recursion_num
-    }
-
-
-@node_monitor_wrapper
-def parse_tool_calling(state: ChatbotState):
-    """executor lambda
-    Args:
-        state (NestUpdateState): _description_
-
-    Returns:
-        _type_: _description_
-    """
-    # parse tool_calls:
-    try:
-        output = _parse_tool_calling(
-            agent_output=state['current_agent_output']
-        )
-        tool_calls = output['tool_calls']
-        send_trace(f"\n\n**tool_calls parsed:** \n{tool_calls}", state["stream"], state["ws_connection_id"], state["enable_trace"])
-        if not state["extra_response"].get("current_agent_intent_type", None):
-            state["extra_response"]["current_agent_intent_type"] = output['tool_calls'][0]["name"]
-       
-        return {
-            "parse_tool_calling_ok": True,
-            "current_tool_calls": tool_calls,
-            "agent_chat_history": [output['agent_message']]
         }
     
-    except (ToolNotExistError,
-             ToolParameterNotExistError,
-             MultipleToolNameError,
-             ToolNotFound
-             ) as e:
-        send_trace(f"\n\n**tool_calls parse failed:** \n{str(e)}", state["stream"], state["ws_connection_id"], state["enable_trace"])
-        return {
-            "parse_tool_calling_ok": False,
-            "agent_chat_history":[
-                e.agent_message,
-                e.error_message
-            ]
-        }
-
-@node_monitor_wrapper
-def tool_execute_lambda(state: ChatbotState):
-    """executor lambda
-    Args:
-        state (NestUpdateState): _description_
-
-    Returns:
-        _type_: _description_
-    """
-    tool_calls = state['current_tool_calls']
-    assert len(tool_calls) == 1, tool_calls
-    tool_call_results = []
-    for tool_call in tool_calls:
-        tool_name = tool_call["name"]
-        tool_kwargs = tool_call['kwargs']
-        # call tool
-        output = invoke_lambda(
-            event_body = {
-                "tool_name":tool_name,
-                "state":state,
-                "kwargs":tool_kwargs
-                },
-            lambda_name="Online_Tool_Execute",
-            lambda_module_path="functions.lambda_tool",
-            handler_name="lambda_handler"   
-        )
-        tool_call_results.append({
-            "name": tool_name,
-            "output": output,
-            "kwargs": tool_call['kwargs'],
-            "model_id": tool_call['model_id']
-        })
+    response = app_agent.invoke({
+        **state,
+        "other_chain_kwargs":other_chain_kwargs
+    })
+    return response
     
-    output = format_tool_call_results(tool_call['model_id'],tool_call_results)
-    send_trace(f'**tool_execute_res:** \n{output["tool_message"]["content"]}')
-    return {
-        "agent_chat_history": [output['tool_message']]
-        }
+    # current_agent_output:dict = invoke_lambda(
+    #     event_body={
+    #         **state,
+    #         "other_chain_kwargs":{
+    #             "goods_info": goods_info,
+    #             "create_time": state['create_time'],
+    #             "current_agent_recursion_num":state['current_agent_recursion_num']
+    #             }
+    #     },
+    #     lambda_name="Online_Agent",
+    #     lambda_module_path="lambda_agent.agent",
+    #     handler_name="lambda_handler"
+    # )
+    # current_agent_recursion_num = state['current_agent_recursion_num'] + 1
+    # send_trace(f"\n\n**current_agent_output:** \n{json.dumps(current_agent_output['agent_output'],ensure_ascii=False,indent=2)}\n\n **current_agent_recursion_num:** {current_agent_recursion_num}", state["stream"], state["ws_connection_id"])
+    # return {
+    #     "current_agent_output": current_agent_output,
+    #     "current_agent_recursion_num": current_agent_recursion_num
+    # }
 
+# @node_monitor_wrapper
+# def tool_execute_lambda(state: ChatbotState):
+#     """executor lambda
+#     Args:
+#         state (NestUpdateState): _description_
 
-@node_monitor_wrapper
-def rag_daily_reception_retriever_lambda(state: ChatbotState):
-    # call retriever
-    retriever_params = state["chatbot_config"]["rag_daily_reception_config"]['retriever_config']
-    retriever_params["query"] = state["query"]
-    output:str = invoke_lambda(
-        event_body=retriever_params,
-        lambda_name="Online_Function_Retriever",
-        lambda_module_path="functions.lambda_retriever.retriever",
-        handler_name="lambda_handler"
-    )
-    contexts = [doc['page_content'] for doc in output['result']['docs']]
-    context = "\n\n".join(contexts)
-    send_trace(f'**rag_goods_exchange_retriever** {context}', state["stream"], state["ws_connection_id"])
-    return {"contexts": contexts}
-
-@node_monitor_wrapper
-def rag_daily_reception_llm_lambda(state:ChatbotState):
-    context = "\n\n".join(state['contexts'])
-    system_prompt = (f"你是安踏的客服助理，正在帮用户解答问题，客户提出的问题大多是属于日常接待类别，你需要按照下面的guidelines进行回复:\n"
-                    "<guidelines>\n"
-                    " - 回复内容需要展现出礼貌。回答内容为一句话，言简意赅。\n"
-                    " - 使用中文回答。\n"
-                    "</guidelines>\n"
-                    "下面列举了一些具体的场景下的回复，你可以结合用户的问题进行参考:\n"
-                    "<context>\n"
-                    f"{context}\n"
-                    "</context>"
-                )
+#     Returns:
+#         _type_: _description_
+#     """
+#     tool_calls = state['current_tool_calls']
+#     assert len(tool_calls) == 1, tool_calls
+#     tool_call_results = []
+#     for tool_call in tool_calls:
+#         tool_name = tool_call["name"]
+#         tool_kwargs = tool_call['kwargs']
+#         # call tool
+#         output = invoke_lambda(
+#             event_body = {
+#                 "tool_name":tool_name,
+#                 "state":state,
+#                 "kwargs":tool_kwargs
+#                 },
+#             lambda_name="Online_Tool_Execute",
+#             lambda_module_path="functions.lambda_tool",
+#             handler_name="lambda_handler"   
+#         )
+#         tool_call_results.append({
+#             "name": tool_name,
+#             "output": output,
+#             "kwargs": tool_call['kwargs'],
+#             "model_id": tool_call['model_id']
+#         })
     
-    output:str = invoke_lambda(
-        lambda_name='Online_LLM_Generate',
-        lambda_module_path="lambda_llm_generate.llm_generate",
-        handler_name='lambda_handler',
-        event_body={
-            "llm_config": {
-                **state['chatbot_config']['rag_daily_reception_config']['llm_config'], 
-                "system_prompt": system_prompt,
-                "intent_type": LLMTaskType.CHAT},
-            "llm_input": {"query": state['query'], "chat_history": state['chat_history']}
-            }
-        )
-    return {"answer": output}
-
-@node_monitor_wrapper
-def rag_goods_exchange_retriever_lambda(state: ChatbotState):
-    # call retriever
-    retriever_params = state["chatbot_config"]["rag_goods_exchange_config"]['retriever_config']
-    retriever_params["query"] = state["query"]
-    output:str = invoke_lambda(
-        event_body=retriever_params,
-        lambda_name="Online_Function_Retriever",
-        lambda_module_path="functions.lambda_retriever.retriever",
-        handler_name="lambda_handler"
-    )
-    contexts = [doc['page_content'] for doc in output['result']['docs']]
-
-    context = "\n\n".join(contexts)
-    send_trace(f'**rag_goods_exchange_retriever** {context}', state["stream"], state["ws_connection_id"])
-    return {"contexts": contexts}
+#     output = format_tool_call_results(tool_call['model_id'],tool_call_results)
+#     send_trace(f'**tool_execute_res:** \n{output["tool_message"]["content"]}')
+#     return {
+#         "agent_chat_history": [output['tool_message']]
+#         }
 
 
-@node_monitor_wrapper
-def rag_goods_exchange_llm_lambda(state:ChatbotState):
-    context = "\n\n".join(state['contexts'])
-    system_prompt = (f"你是安踏的客服助理，正在帮用户解答问题，客户提出的问题大多是属于商品退换货范畴，你需要按照下面的guidelines进行回复:\n"
-                    "<guidelines>\n"
-                    " - 回复内容需要展现出礼貌。回答内容为一句话，言简意赅。\n"
-                    " - 使用中文回答。\n"
-                    "</guidelines>\n"
-                    "下面列举了一些具体的场景下的回复，你可以结合用户的问题进行参考回答:\n"
-                    "<context>\n"
-                    f"{context}\n"
-                    "</context>\n"
-                )
+# @node_monitor_wrapper
+# def rag_daily_reception_retriever_lambda(state: ChatbotState):
+#     # call retriever
+#     retriever_params = state["chatbot_config"]["rag_daily_reception_config"]['retriever_config']
+#     retriever_params["query"] = state["query"]
+#     output:str = invoke_lambda(
+#         event_body=retriever_params,
+#         lambda_name="Online_Function_Retriever",
+#         lambda_module_path="functions.lambda_retriever.retriever",
+#         handler_name="lambda_handler"
+#     )
+#     contexts = [doc['page_content'] for doc in output['result']['docs']]
+#     context = "\n\n".join(contexts)
+#     send_trace(f'**rag_goods_exchange_retriever** {context}', state["stream"], state["ws_connection_id"])
+#     return {"contexts": contexts}
+
+# @node_monitor_wrapper
+# def rag_daily_reception_llm_lambda(state:ChatbotState):
+#     context = "\n\n".join(state['contexts'])
+#     system_prompt = (f"你是安踏的客服助理，正在帮用户解答问题，客户提出的问题大多是属于日常接待类别，你需要按照下面的guidelines进行回复:\n"
+#                     "<guidelines>\n"
+#                     " - 回复内容需要展现出礼貌。回答内容为一句话，言简意赅。\n"
+#                     " - 使用中文回答。\n"
+#                     "</guidelines>\n"
+#                     "下面列举了一些具体的场景下的回复，你可以结合用户的问题进行参考:\n"
+#                     "<context>\n"
+#                     f"{context}\n"
+#                     "</context>"
+#                 )
     
-    output:str = invoke_lambda(
-        lambda_name='Online_LLM_Generate',
-        lambda_module_path="lambda_llm_generate.llm_generate",
-        handler_name='lambda_handler',
-        event_body={
-            "llm_config": {
-                **state['chatbot_config']['rag_goods_exchange_config']['llm_config'],
-                "system_prompt":system_prompt,
-                "intent_type": LLMTaskType.CHAT},
-            "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
-            }
-        )
-    return {"answer": output}
+#     output:str = invoke_lambda(
+#         lambda_name='Online_LLM_Generate',
+#         lambda_module_path="lambda_llm_generate.llm_generate",
+#         handler_name='lambda_handler',
+#         event_body={
+#             "llm_config": {
+#                 **state['chatbot_config']['rag_daily_reception_config']['llm_config'], 
+#                 "system_prompt": system_prompt,
+#                 "intent_type": LLMTaskType.CHAT},
+#             "llm_input": {"query": state['query'], "chat_history": state['chat_history']}
+#             }
+#         )
+#     return {"answer": output}
 
-@node_monitor_wrapper
-def rag_product_aftersales_retriever_lambda(state: ChatbotState):
-    # call retriever
-    recent_tool_calling:list[dict] = state['current_tool_calls'][0]
-    if "shop" in recent_tool_calling['kwargs'] and recent_tool_calling['kwargs']['shop'] != "tianmao":
-        contexts = ["顾客不是在天猫购买的商品，请他咨询其他商家"]
-        return {"contexts": contexts}
-    retriever_params = state["chatbot_config"]["rag_product_aftersales_config"]["retriever_config"]
-    retriever_params["query"] = state["query"]
-    output:str = invoke_lambda(
-        event_body=retriever_params,
-        lambda_name="Online_Function_Retriever",
-        lambda_module_path="functions.lambda_retriever.retriever",
-        handler_name="lambda_handler"
-    )
-    contexts = [doc['page_content'] for doc in output['result']['docs']]
+# @node_monitor_wrapper
+# def rag_goods_exchange_retriever_lambda(state: ChatbotState):
+#     # call retriever
+#     retriever_params = state["chatbot_config"]["rag_goods_exchange_config"]['retriever_config']
+#     retriever_params["query"] = state["query"]
+#     output:str = invoke_lambda(
+#         event_body=retriever_params,
+#         lambda_name="Online_Function_Retriever",
+#         lambda_module_path="functions.lambda_retriever.retriever",
+#         handler_name="lambda_handler"
+#     )
+#     contexts = [doc['page_content'] for doc in output['result']['docs']]
 
-    context = "\n\n".join(contexts)
-    send_trace(f'**rag_product_aftersales_retriever** {context}', state["stream"], state["ws_connection_id"])
-    return {"contexts": contexts}
+#     context = "\n\n".join(contexts)
+#     send_trace(f'**rag_goods_exchange_retriever** {context}', state["stream"], state["ws_connection_id"])
+#     return {"contexts": contexts}
 
-@node_monitor_wrapper
-def rag_product_aftersales_llm_lambda(state:ChatbotState):
-    create_time = state.get('create_time', None)
-    goods_id = state.get('chatbot_config').get('goods_id', 757492962957)
-    try:
-        create_datetime_object = datetime.strptime(create_time, '%Y-%m-%d %H:%M:%S.%f')
-    except Exception as e:
-        create_datetime_object = datetime.now()
-        print(f"create_time: {create_time} is not valid, use current time instead.")
-    create_time_str = create_datetime_object.strftime('%Y-%m-%d')
-    # TODO: fix received time format
-    received_time = order_dict.get(str(goods_id), {}).get("received_time", "2023/9/129:03:13")
-    order_time = " ".join([received_time[:9], received_time[9:]])
-    try:
-        order_date_str = datetime.strptime(order_time, '%Y/%m/%d %H:%M:%S').strftime('%Y-%m-%d')
-        receive_elapsed_days = (create_datetime_object - datetime.strptime(order_date_str, '%Y-%m-%d')).days
-        receive_elapsed_months = receive_elapsed_days // 30
-    except Exception as e:
-        order_date_str = "2023-9-12"
-        receive_elapsed_months = 6
-    context = "\n\n".join(state['contexts'])
-    system_prompt = (f"你是安踏的客服助理，正在帮消费者解答问题，消费者提出的问题大多是属于商品的质量和物流规则。context列举了一些可能有关的具体场景及回复，你可以进行参考:\n"
-                    f"客户咨询的问题所对应的订单日期为{order_date_str}。\n"
-                    f"当前时间{create_time_str}\n"
-                    f"客户收到商品已经超过{receive_elapsed_months}个月\n"
-                    "<context>\n"
-                    f"{context}\n"
-                    "</context>\n"
-                    "你需要按照下面的guidelines对消费者的问题进行回答:\n"
-                    "<guidelines>\n"
-                    " - 回答内容为一句话，言简意赅。\n"
-                    " - 如果问题与context内容不相关，就不要采用。\n"
-                    " - 消费者的问题里面可能包含口语化的表达，比如鞋子开胶的意思是用胶黏合的鞋体裂开。这和胶丝遗留没有关系\n"
-                    " - 洗涤后出现问题也属于质量问题\n"
-                    " - 消费者的回复不够清晰的时候，直接回复: 不知道刚才给您的建议是否有帮助？。不要有额外补充\n"
-                    " - 如果客户问到质量相关问题，请根据前面的订单信息和三包规则，确定是否超出三包期限，如果超出三包期限请告知消费者无法处理，如果在三包期限内请按照三包要求处理，并安抚。\n"
-                    "</guidelines>\n"
-                    )
-    # print('llm config',state['chatbot_config']['rag_product_aftersales_config']['llm_config'])
-    output:str = invoke_lambda(
-        lambda_name='Online_LLM_Generate',
-        lambda_module_path="lambda_llm_generate.llm_generate",
-        handler_name='lambda_handler',
-        event_body={
-            "llm_config": {
-                **state['chatbot_config']['rag_product_aftersales_config']['llm_config'], 
-                "system_prompt":system_prompt,
-                "intent_type": LLMTaskType.CHAT
-            },
-            "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
-            }
-        )
-    return {"answer": output}
 
-@node_monitor_wrapper
-def rag_customer_complain_retriever_lambda(state: ChatbotState):
-    # call retriever
-    retriever_params = state["chatbot_config"]["rag_customer_complain_config"]["retriever_config"]
-    retriever_params["query"] = state["query"]
-    output:str = invoke_lambda(
-        event_body=retriever_params,
-        lambda_name="Online_Function_Retriever",
-        lambda_module_path="functions.lambda_retriever.retriever",
-        handler_name="lambda_handler"
-    )
-    contexts = [doc['page_content'] for doc in output['result']['docs']]
+# @node_monitor_wrapper
+# def rag_goods_exchange_llm_lambda(state:ChatbotState):
+#     context = "\n\n".join(state['contexts'])
+#     system_prompt = (f"你是安踏的客服助理，正在帮用户解答问题，客户提出的问题大多是属于商品退换货范畴，你需要按照下面的guidelines进行回复:\n"
+#                     "<guidelines>\n"
+#                     " - 回复内容需要展现出礼貌。回答内容为一句话，言简意赅。\n"
+#                     " - 使用中文回答。\n"
+#                     "</guidelines>\n"
+#                     "下面列举了一些具体的场景下的回复，你可以结合用户的问题进行参考回答:\n"
+#                     "<context>\n"
+#                     f"{context}\n"
+#                     "</context>\n"
+#                 )
+    
+#     output:str = invoke_lambda(
+#         lambda_name='Online_LLM_Generate',
+#         lambda_module_path="lambda_llm_generate.llm_generate",
+#         handler_name='lambda_handler',
+#         event_body={
+#             "llm_config": {
+#                 **state['chatbot_config']['rag_goods_exchange_config']['llm_config'],
+#                 "system_prompt":system_prompt,
+#                 "intent_type": LLMTaskType.CHAT},
+#             "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
+#             }
+#         )
+#     return {"answer": output}
 
-    context = "\n\n".join(contexts)
-    send_trace(f'**rag_customer_complain_retriever** {context}', state["stream"], state["ws_connection_id"])
-    return {"contexts": contexts}
+# @node_monitor_wrapper
+# def rag_product_aftersales_retriever_lambda(state: ChatbotState):
+#     # call retriever
+#     recent_tool_calling:list[dict] = state['current_tool_calls'][0]
+#     if "shop" in recent_tool_calling['kwargs'] and recent_tool_calling['kwargs']['shop'] != "tianmao":
+#         contexts = ["顾客不是在天猫购买的商品，请他咨询其他商家"]
+#         return {"contexts": contexts}
+#     retriever_params = state["chatbot_config"]["rag_product_aftersales_config"]["retriever_config"]
+#     retriever_params["query"] = state["query"]
+#     output:str = invoke_lambda(
+#         event_body=retriever_params,
+#         lambda_name="Online_Function_Retriever",
+#         lambda_module_path="functions.lambda_retriever.retriever",
+#         handler_name="lambda_handler"
+#     )
+#     contexts = [doc['page_content'] for doc in output['result']['docs']]
 
-@node_monitor_wrapper
-def rag_customer_complain_llm_lambda(state:ChatbotState):
-    context = "\n\n".join(state['contexts'])
-    # prompt = dedent(f"""你是安踏的客服助理，正在处理有关于客户抱怨的问题，这些问题有关于商品质量等方面，需要你按照下面的guidelines进行回复:
-    system_prompt = ("你是安踏的客服助理，正在处理有关于消费者抱怨的问题。context列举了一些可能和客户问题有关的具体场景及回复，你可以进行参考:\n"
-                    "<context>\n"
-                    f"{context}\n"
-                    "</context>\n"
-                    "需要你按照下面的guidelines进行回复:\n"
-                    "<guidelines>\n"
-                    " - 回答内容为一句话，言简意赅。\n"
-                    " - 尽量安抚客户情绪。\n"
-                    " - 直接回答，不要说\"亲爱的顾客，您好\"\n"
-                    "</guidelines>\n"
-                    )
-    output:str = invoke_lambda(
-        lambda_name='Online_LLM_Generate',
-        lambda_module_path="lambda_llm_generate.llm_generate",
-        handler_name='lambda_handler',
-        event_body={
-            "llm_config": {
-                **state['chatbot_config']['rag_customer_complain_config']['llm_config'],
-                "system_prompt":system_prompt,
-                "intent_type": LLMTaskType.CHAT},
-            "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
-            }
-        )
-    return {"answer": output}
+#     context = "\n\n".join(contexts)
+#     send_trace(f'**rag_product_aftersales_retriever** {context}', state["stream"], state["ws_connection_id"])
+#     return {"contexts": contexts}
 
-@node_monitor_wrapper
-def rag_promotion_retriever_lambda(state: ChatbotState):
-    # call retriever
-    retriever_params = state["chatbot_config"]["rag_promotion_config"]["retriever_config"]
-    retriever_params["query"] = state["query"]
-    output:str = invoke_lambda(
-        event_body=retriever_params,
-        lambda_name="Online_Function_Retriever",
-        lambda_module_path="functions.lambda_retriever.retriever",
-        handler_name="lambda_handler"
-    )
-    contexts = [doc['page_content'] for doc in output['result']['docs']]
+# @node_monitor_wrapper
+# def rag_product_aftersales_llm_lambda(state:ChatbotState):
+#     create_time = state.get('create_time', None)
+#     goods_id = state.get('chatbot_config').get('goods_id', 757492962957)
+#     try:
+#         create_datetime_object = datetime.strptime(create_time, '%Y-%m-%d %H:%M:%S.%f')
+#     except Exception as e:
+#         create_datetime_object = datetime.now()
+#         print(f"create_time: {create_time} is not valid, use current time instead.")
+#     create_time_str = create_datetime_object.strftime('%Y-%m-%d')
+#     # TODO: fix received time format
+#     received_time = order_dict.get(str(goods_id), {}).get("received_time", "2023/9/129:03:13")
+#     order_time = " ".join([received_time[:9], received_time[9:]])
+#     try:
+#         order_date_str = datetime.strptime(order_time, '%Y/%m/%d %H:%M:%S').strftime('%Y-%m-%d')
+#         receive_elapsed_days = (create_datetime_object - datetime.strptime(order_date_str, '%Y-%m-%d')).days
+#         receive_elapsed_months = receive_elapsed_days // 30
+#     except Exception as e:
+#         order_date_str = "2023-9-12"
+#         receive_elapsed_months = 6
+#     context = "\n\n".join(state['contexts'])
+#     system_prompt = (f"你是安踏的客服助理，正在帮消费者解答问题，消费者提出的问题大多是属于商品的质量和物流规则。context列举了一些可能有关的具体场景及回复，你可以进行参考:\n"
+#                     f"客户咨询的问题所对应的订单日期为{order_date_str}。\n"
+#                     f"当前时间{create_time_str}\n"
+#                     f"客户收到商品已经超过{receive_elapsed_months}个月\n"
+#                     "<context>\n"
+#                     f"{context}\n"
+#                     "</context>\n"
+#                     "你需要按照下面的guidelines对消费者的问题进行回答:\n"
+#                     "<guidelines>\n"
+#                     " - 回答内容为一句话，言简意赅。\n"
+#                     " - 如果问题与context内容不相关，就不要采用。\n"
+#                     " - 消费者的问题里面可能包含口语化的表达，比如鞋子开胶的意思是用胶黏合的鞋体裂开。这和胶丝遗留没有关系\n"
+#                     " - 洗涤后出现问题也属于质量问题\n"
+#                     " - 消费者的回复不够清晰的时候，直接回复: 不知道刚才给您的建议是否有帮助？。不要有额外补充\n"
+#                     " - 如果客户问到质量相关问题，请根据前面的订单信息和三包规则，确定是否超出三包期限，如果超出三包期限请告知消费者无法处理，如果在三包期限内请按照三包要求处理，并安抚。\n"
+#                     "</guidelines>\n"
+#                     )
+#     # print('llm config',state['chatbot_config']['rag_product_aftersales_config']['llm_config'])
+#     output:str = invoke_lambda(
+#         lambda_name='Online_LLM_Generate',
+#         lambda_module_path="lambda_llm_generate.llm_generate",
+#         handler_name='lambda_handler',
+#         event_body={
+#             "llm_config": {
+#                 **state['chatbot_config']['rag_product_aftersales_config']['llm_config'], 
+#                 "system_prompt":system_prompt,
+#                 "intent_type": LLMTaskType.CHAT
+#             },
+#             "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
+#             }
+#         )
+#     return {"answer": output}
 
-    context = "\n\n".join(contexts)
-    send_trace(f'**rag_promotion_retriever** {context}', state["stream"], state["ws_connection_id"])
-    return {"contexts": contexts}
+# @node_monitor_wrapper
+# def rag_customer_complain_retriever_lambda(state: ChatbotState):
+#     # call retriever
+#     retriever_params = state["chatbot_config"]["rag_customer_complain_config"]["retriever_config"]
+#     retriever_params["query"] = state["query"]
+#     output:str = invoke_lambda(
+#         event_body=retriever_params,
+#         lambda_name="Online_Function_Retriever",
+#         lambda_module_path="functions.lambda_retriever.retriever",
+#         handler_name="lambda_handler"
+#     )
+#     contexts = [doc['page_content'] for doc in output['result']['docs']]
 
-@node_monitor_wrapper
-def rag_promotion_llm_lambda(state:ChatbotState):
-    context = "\n\n".join(state['contexts'])
+#     context = "\n\n".join(contexts)
+#     send_trace(f'**rag_customer_complain_retriever** {context}', state["stream"], state["ws_connection_id"])
+#     return {"contexts": contexts}
 
-    system_prompt = ("你是安踏的客服助理，正在帮消费者解答有关于商品促销的问题，这些问题是有关于积分、奖品、奖励等方面。\n"
-                     "context列举了一些可能有关的具体场景及回复，你可以进行参考:\n"
-                    f"<context>\n{context}\n</context>\n"
-                    "你需要按照下面的guidelines对消费者的问题进行回答:\n"
-                    "<guidelines>\n"
-                    " - 回答内容为一句话，言简意赅。\n"
-                    " - 如果问题与context内容不相关，就不要采用。\n"
-                    " - 使用中文进行回答。\n"
-                    "</guidelines>"
-    )
+# @node_monitor_wrapper
+# def rag_customer_complain_llm_lambda(state:ChatbotState):
+#     context = "\n\n".join(state['contexts'])
+#     # prompt = dedent(f"""你是安踏的客服助理，正在处理有关于客户抱怨的问题，这些问题有关于商品质量等方面，需要你按照下面的guidelines进行回复:
+#     system_prompt = ("你是安踏的客服助理，正在处理有关于消费者抱怨的问题。context列举了一些可能和客户问题有关的具体场景及回复，你可以进行参考:\n"
+#                     "<context>\n"
+#                     f"{context}\n"
+#                     "</context>\n"
+#                     "需要你按照下面的guidelines进行回复:\n"
+#                     "<guidelines>\n"
+#                     " - 回答内容为一句话，言简意赅。\n"
+#                     " - 尽量安抚客户情绪。\n"
+#                     " - 直接回答，不要说\"亲爱的顾客，您好\"\n"
+#                     "</guidelines>\n"
+#                     )
+#     output:str = invoke_lambda(
+#         lambda_name='Online_LLM_Generate',
+#         lambda_module_path="lambda_llm_generate.llm_generate",
+#         handler_name='lambda_handler',
+#         event_body={
+#             "llm_config": {
+#                 **state['chatbot_config']['rag_customer_complain_config']['llm_config'],
+#                 "system_prompt":system_prompt,
+#                 "intent_type": LLMTaskType.CHAT},
+#             "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
+#             }
+#         )
+#     return {"answer": output}
+
+# @node_monitor_wrapper
+# def rag_promotion_retriever_lambda(state: ChatbotState):
+#     # call retriever
+#     retriever_params = state["chatbot_config"]["rag_promotion_config"]["retriever_config"]
+#     retriever_params["query"] = state["query"]
+#     output:str = invoke_lambda(
+#         event_body=retriever_params,
+#         lambda_name="Online_Function_Retriever",
+#         lambda_module_path="functions.lambda_retriever.retriever",
+#         handler_name="lambda_handler"
+#     )
+#     contexts = [doc['page_content'] for doc in output['result']['docs']]
+
+#     context = "\n\n".join(contexts)
+#     send_trace(f'**rag_promotion_retriever** {context}', state["stream"], state["ws_connection_id"])
+#     return {"contexts": contexts}
+
+# @node_monitor_wrapper
+# def rag_promotion_llm_lambda(state:ChatbotState):
+#     context = "\n\n".join(state['contexts'])
+
+#     system_prompt = ("你是安踏的客服助理，正在帮消费者解答有关于商品促销的问题，这些问题是有关于积分、奖品、奖励等方面。\n"
+#                      "context列举了一些可能有关的具体场景及回复，你可以进行参考:\n"
+#                     f"<context>\n{context}\n</context>\n"
+#                     "你需要按照下面的guidelines对消费者的问题进行回答:\n"
+#                     "<guidelines>\n"
+#                     " - 回答内容为一句话，言简意赅。\n"
+#                     " - 如果问题与context内容不相关，就不要采用。\n"
+#                     " - 使用中文进行回答。\n"
+#                     "</guidelines>"
+#     )
    
-    output:str = invoke_lambda(
-        lambda_name='Online_LLM_Generate',
-        lambda_module_path="lambda_llm_generate.llm_generate",
-        handler_name='lambda_handler',
-        event_body={
-            "llm_config": {
-                **state['chatbot_config']['rag_promotion_config']['llm_config'],
-                "system_prompt":system_prompt,
-                "intent_type": LLMTaskType.CHAT},
-            "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
-            }
-        )
-    return {"answer": output}
+#     output:str = invoke_lambda(
+#         lambda_name='Online_LLM_Generate',
+#         lambda_module_path="lambda_llm_generate.llm_generate",
+#         handler_name='lambda_handler',
+#         event_body={
+#             "llm_config": {
+#                 **state['chatbot_config']['rag_promotion_config']['llm_config'],
+#                 "system_prompt":system_prompt,
+#                 "intent_type": LLMTaskType.CHAT},
+#             "llm_input": { "query": state['query'], "chat_history": state['chat_history']}
+#             }
+#         )
+#     return {"answer": output}
 
 
 
@@ -562,87 +545,89 @@ def final_rag_llm_lambda(state:ChatbotState):
     return {"answer": output}
 
 
-def transfer_reply(state:ChatbotState):
-    return {"answer": "您好,我是安踏官方客服,很高兴为您服务。请问您有什么需要帮助的吗?"}
+# def transfer_reply(state:ChatbotState):
+#     return {"answer": "您好,我是安踏官方客服,很高兴为您服务。请问您有什么需要帮助的吗?"}
 
 
-def give_rhetorical_question(state:ChatbotState):
-    recent_tool_calling:list[dict] = state['current_tool_calls'][0]
-    return {"answer": recent_tool_calling['kwargs']['question']}
+# def give_rhetorical_question(state:ChatbotState):
+#     recent_tool_calling:list[dict] = state['current_tool_calls'][0]
+#     return {"answer": recent_tool_calling['kwargs']['question']}
 
 
-def give_final_response(state:ChatbotState):
-    recent_tool_calling:list[dict] = state['current_tool_calls'][0]
-    return {"answer": recent_tool_calling['kwargs']['response']}
+# def give_final_response(state:ChatbotState):
+#     recent_tool_calling:list[dict] = state['current_tool_calls'][0]
+#     return {"answer": recent_tool_calling['kwargs']['response']}
 
-def rule_url_reply(state:ChatbotState):
-    state["extra_response"]["current_agent_intent_type"] = "rule reply"
-    if state['query'].endswith(('.jpg','.png')):
-        answer = random.choice([
-            "收到，亲。请问我们可以怎么为您效劳呢？",
-            "您好，请问有什么需要帮助的吗？"
-        ])
-        return {"answer": answer}
-    # product information
-    r = re.findall(r"item.htm\?id=(.*)",state['query'])
-    if r:
-        goods_id = r[0]
-    else:
-        goods_id = 0
-    if goods_id in goods_dict:
-        # call llm to make summary of goods info
-        human_goods_info = state['human_goods_info']
-        output = f"您好，该商品的特点是:\n{human_goods_info}"
-        if human_goods_info:
-            system_prompt = (f"你是安踏的客服助理，当前用户对下面的商品感兴趣:\n"
-                        f"<{goods_info_tag}>\n{human_goods_info}\n</{goods_info_tag}>\n"
-                        "请你结合商品的基础信息，特别是卖点信息返回一句推荐语。"
-                    )
-            output:str = invoke_lambda(
-                lambda_name='Online_LLM_Generate',
-                lambda_module_path="lambda_llm_generate.llm_generate",
-                handler_name='lambda_handler',
-                event_body={
-                    "llm_config": {
-                        **state['chatbot_config']['rag_daily_reception_config']['llm_config'], 
-                        "system_prompt": system_prompt,
-                        "intent_type": LLMTaskType.CHAT},
-                    "llm_input": {"query": state['query'], "chat_history": state['chat_history']}
-                        }
-                    )
+# def rule_url_reply(state:ChatbotState):
+#     state["extra_response"]["current_agent_intent_type"] = "rule reply"
+#     if state['query'].endswith(('.jpg','.png')):
+#         answer = random.choice([
+#             "收到，亲。请问我们可以怎么为您效劳呢？",
+#             "您好，请问有什么需要帮助的吗？"
+#         ])
+#         return {"answer": answer}
+#     # product information
+#     r = re.findall(r"item.htm\?id=(.*)",state['query'])
+#     if r:
+#         goods_id = r[0]
+#     else:
+#         goods_id = 0
+#     if goods_id in goods_dict:
+#         # call llm to make summary of goods info
+#         human_goods_info = state['human_goods_info']
+#         output = f"您好，该商品的特点是:\n{human_goods_info}"
+#         if human_goods_info:
+#             system_prompt = (f"你是安踏的客服助理，当前用户对下面的商品感兴趣:\n"
+#                         f"<{goods_info_tag}>\n{human_goods_info}\n</{goods_info_tag}>\n"
+#                         "请你结合商品的基础信息，特别是卖点信息返回一句推荐语。"
+#                     )
+#             output:str = invoke_lambda(
+#                 lambda_name='Online_LLM_Generate',
+#                 lambda_module_path="lambda_llm_generate.llm_generate",
+#                 handler_name='lambda_handler',
+#                 event_body={
+#                     "llm_config": {
+#                         **state['chatbot_config']['rag_daily_reception_config']['llm_config'], 
+#                         "system_prompt": system_prompt,
+#                         "intent_type": LLMTaskType.CHAT},
+#                     "llm_input": {"query": state['query'], "chat_history": state['chat_history']}
+#                         }
+#                     )
          
-        return {"answer":output}
+#         return {"answer":output}
     
-    return {"answer":"您好"}
+#     return {"answer":"您好"}
 
 def rule_number_reply(state:ChatbotState):
     state["extra_response"]["current_agent_intent_type"] = "rule reply"
     return {"answer":"收到订单信息"}
 
 
+def final_results_preparation(state: ChatbotState):
+    return {"answer": state['answer']}
 
-@node_monitor_wrapper
-def tools_choose_and_results_generation(state: ChatbotState):
-    # check once tool calling
-    current_agent_output:dict = invoke_lambda(
-        event_body={
-            **state,
-            # "other_chain_kwargs": {"system_prompt": get_common_system_prompt()}
-            },
-        lambda_name="Online_Agent",
-        lambda_module_path="lambda_agent.agent",
-        handler_name="lambda_handler",
+# @node_monitor_wrapper
+# def tools_choose_and_results_generation(state: ChatbotState):
+#     # check once tool calling
+#     current_agent_output:dict = invoke_lambda(
+#         event_body={
+#             **state,
+#             # "other_chain_kwargs": {"system_prompt": get_common_system_prompt()}
+#             },
+#         lambda_name="Online_Agent",
+#         lambda_module_path="lambda_agent.agent",
+#         handler_name="lambda_handler",
    
-    )
-    current_agent_recursion_num = state['current_agent_recursion_num'] + 1
-    agent_recursion_validation = state['current_agent_recursion_num'] < state['agent_recursion_limit']
+#     )
+#     current_agent_recursion_num = state['current_agent_recursion_num'] + 1
+#     agent_recursion_validation = state['current_agent_recursion_num'] < state['agent_recursion_limit']
 
-    send_trace(f"\n\n**current_agent_output:** \n{json.dumps(current_agent_output['agent_output'],ensure_ascii=False,indent=2)}\n\n **current_agent_recursion_num:** {current_agent_recursion_num}", state["stream"], state["ws_connection_id"])
-    return {
-        "current_agent_output": current_agent_output,
-        "current_agent_recursion_num": current_agent_recursion_num,
-        "agent_recursion_validation": agent_recursion_validation
-    }
+#     send_trace(f"\n\n**current_agent_output:** \n{json.dumps(current_agent_output['agent_output'],ensure_ascii=False,indent=2)}\n\n **current_agent_recursion_num:** {current_agent_recursion_num}", state["stream"], state["ws_connection_id"])
+#     return {
+#         "current_agent_output": current_agent_output,
+#         "current_agent_recursion_num": current_agent_recursion_num,
+#         "agent_recursion_validation": agent_recursion_validation
+#     }
 
 
 ################
@@ -666,59 +651,71 @@ def query_route(state:dict):
 def intent_route(state:dict):
     return state['intent_type']
 
-def agent_route(state:dict):
-    parse_tool_calling_ok = state['parse_tool_calling_ok']
-    if not parse_tool_calling_ok:
-        if state['current_agent_recursion_num'] >= state['agent_recursion_limit']:
-            send_trace(f"Reach the agent recursion limit: {state['agent_recursion_limit']}, route to final rag")
-            return 'final rag'
-        return 'invalid tool calling'
+# def agent_route(state:dict):
+#     parse_tool_calling_ok = state['parse_tool_calling_ok']
+#     if not parse_tool_calling_ok:
+#         if state['current_agent_recursion_num'] >= state['agent_recursion_limit']:
+#             send_trace(f"Reach the agent recursion limit: {state['agent_recursion_limit']}, route to final rag")
+#             return 'final rag'
+#         return 'invalid tool calling'
     
-    recent_tool_calls:list[dict] = state['current_tool_calls']
+#     recent_tool_calls:list[dict] = state['current_tool_calls']
     
-    recent_tool_call = recent_tool_calls[0]
+#     recent_tool_call = recent_tool_calls[0]
 
-    recent_tool_name = recent_tool_call['name']
+#     recent_tool_name = recent_tool_call['name']
 
-    if recent_tool_name in ['comfort', 'transfer']:
-        return recent_tool_name
+#     if recent_tool_name in ['comfort', 'transfer']:
+#         return recent_tool_name
     
-    if recent_tool_call['name'] == "give_rhetorical_question":
-        return "rhetorical question"
+#     if recent_tool_call['name'] == "give_rhetorical_question":
+#         return "rhetorical question"
     
-    if recent_tool_call['name'] == "goods_exchange":
-        return "goods exchange"
+#     if recent_tool_call['name'] == "goods_exchange":
+#         return "goods exchange"
     
-    if recent_tool_call['name'] == "daily_reception":
-        return "daily reception"
+#     if recent_tool_call['name'] == "daily_reception":
+#         return "daily reception"
 
-    if recent_tool_call['name'] == "rule_response":
-        return "rule response"
+#     if recent_tool_call['name'] == "rule_response":
+#         return "rule response"
 
-    if recent_tool_call['name'] == 'product_logistics':
-        return "product aftersales"
+#     if recent_tool_call['name'] == 'product_logistics':
+#         return "product aftersales"
 
-    if recent_tool_call['name'] == 'product_quality':
-        return "product aftersales"
+#     if recent_tool_call['name'] == 'product_quality':
+#         return "product aftersales"
 
-    if recent_tool_call['name'] == 'goods_storage':
-        return "product aftersales"
+#     if recent_tool_call['name'] == 'goods_storage':
+#         return "product aftersales"
 
-    if recent_tool_call['name'] == 'customer_complain':
-        return "customer complain"
+#     if recent_tool_call['name'] == 'customer_complain':
+#         return "customer complain"
 
-    if recent_tool_call['name'] == 'promotion':
-        return "promotion"
+#     if recent_tool_call['name'] == 'promotion':
+#         return "promotion"
     
-    if recent_tool_call['name'] == "give_final_response":
-        return "give final response"
+#     if recent_tool_call['name'] == "give_final_response":
+#         return "give final response"
 
-    if state['current_agent_recursion_num'] >= state['agent_recursion_limit']:
-        send_trace(f"Reach the agent recursion limit: {state['agent_recursion_limit']}, route to final rag")
-        return 'final rag'
+#     if state['current_agent_recursion_num'] >= state['agent_recursion_limit']:
+#         send_trace(f"Reach the agent recursion limit: {state['agent_recursion_limit']}, route to final rag")
+#         return 'final rag'
 
-    return "continue"
+#     return "continue"
 
+
+def agent_route(state: dict):
+    if state.get("is_current_tool_calling_once",False):
+        return "no need tool calling"
+    
+    state["agent_recursion_validation"] = state['current_agent_recursion_num'] < state['agent_recursion_limit']
+
+    if state["agent_recursion_validation"]:
+        return "valid tool calling"
+    else:
+        # TODO give final strategy
+        raise 'final rag'
 
      
 #############################
@@ -727,90 +724,61 @@ def agent_route(state:dict):
 
 app_agent = None
 
-def build_agent_graph():
-    def _results_evaluation_route(state: dict):
-        #TODO: pass no need tool calling or valid tool calling?
-        if state["agent_recursion_validation"] and not state["parse_tool_calling_ok"]:
-            return "invalid tool calling"
-        return "continue"
-
-    workflow = StateGraph(ChatbotState)
-    workflow.add_node("tools_choose_and_results_generation", tools_choose_and_results_generation)
-    workflow.add_node("results_evaluation", results_evaluation)
-
-    # add all edges
-    workflow.set_entry_point("tools_choose_and_results_generation")
-    workflow.add_edge("tools_choose_and_results_generation","results_evaluation")
-
-    # add conditional edges
-    # the results of agent planning will be evaluated and decide next step:
-    # 1. invalid tool calling: if agent makes clear mistakes, like wrong tool names or format, it will be forced to plan again
-    # 2. valid tool calling: the agent chooses the valid tools
-    workflow.add_conditional_edges(
-        "results_evaluation",
-        _results_evaluation_route,
-        {
-            "invalid tool calling": "tools_choose_and_results_generation",
-            "continue": END,
-        }
-    )
-    app = workflow.compile()
-    return app
-
-
 def build_graph():
     workflow = StateGraph(ChatbotState)
     # add all nodes
-    workflow.add_node("query_preprocess_lambda", query_preprocess_lambda)
-    workflow.add_node("intention_detection_lambda", intention_detection_lambda)
-    workflow.add_node("agent_lambda", agent_lambda)
-    workflow.add_node("tool_execute_lambda", tool_execute_lambda)
-    workflow.add_node("transfer_reply", transfer_reply)
-    workflow.add_node("give_rhetorical_question",give_rhetorical_question)
-    workflow.add_node("give_final_response",give_final_response)
+    workflow.add_node("query_preprocess", query_preprocess)
+    workflow.add_node("intention_detection", intention_detection)
+    workflow.add_node("agent", agent)
+    workflow.add_node("tool_execute", tool_execution)
+    # workflow.add_node("transfer_reply", transfer_reply)
+    # workflow.add_node("give_rhetorical_question",give_rhetorical_question)
+    # workflow.add_node("give_final_response",give_final_response)
     # workflow.add_node("give_response_wo_tool",give_response_without_any_tool)
-    workflow.add_node("parse_tool_calling",parse_tool_calling)
+    # workflow.add_node("parse_tool_calling",parse_tool_calling)
     # 
-    workflow.add_node("rag_daily_reception_retriever",rag_daily_reception_retriever_lambda)
-    workflow.add_node("rag_daily_reception_llm",rag_daily_reception_llm_lambda)
-    workflow.add_node("rag_goods_exchange_retriever",rag_goods_exchange_retriever_lambda)
-    workflow.add_node("rag_goods_exchange_llm",rag_goods_exchange_llm_lambda)
-    workflow.add_node("rag_product_aftersales_retriever",rag_product_aftersales_retriever_lambda)
-    workflow.add_node("rag_product_aftersales_llm",rag_product_aftersales_llm_lambda)
-    workflow.add_node("rag_customer_complain_retriever",rag_customer_complain_retriever_lambda)
-    workflow.add_node("rag_customer_complain_llm",rag_customer_complain_llm_lambda)
-    workflow.add_node("rule_url_reply",rule_url_reply)
+    # workflow.add_node("rag_daily_reception_retriever",rag_daily_reception_retriever_lambda)
+    # workflow.add_node("rag_daily_reception_llm",rag_daily_reception_llm_lambda)
+    # workflow.add_node("rag_goods_exchange_retriever",rag_goods_exchange_retriever_lambda)
+    # workflow.add_node("rag_goods_exchange_llm",rag_goods_exchange_llm_lambda)
+    # workflow.add_node("rag_product_aftersales_retriever",rag_product_aftersales_retriever_lambda)
+    # workflow.add_node("rag_product_aftersales_llm",rag_product_aftersales_llm_lambda)
+    # workflow.add_node("rag_customer_complain_retriever",rag_customer_complain_retriever_lambda)
+    # workflow.add_node("rag_customer_complain_llm",rag_customer_complain_llm_lambda)
+    # workflow.add_node("rule_url_reply",rule_url_reply)
     workflow.add_node("rule_number_reply",rule_number_reply)
-    workflow.add_node("rag_promotion_retriever",rag_promotion_retriever_lambda)
-    workflow.add_node("rag_promotion_llm",rag_promotion_llm_lambda)
+    # workflow.add_node("rag_promotion_retriever",rag_promotion_retriever_lambda)
+    # workflow.add_node("rag_promotion_llm",rag_promotion_llm_lambda)
     workflow.add_node("final_rag_retriever",final_rag_retriever_lambda)
     workflow.add_node("final_rag_llm",final_rag_llm_lambda)
 
+    workflow.add_node("final_results_preparation", final_results_preparation)
+
     # add all edges
-    workflow.set_entry_point("query_preprocess_lambda")
+    workflow.set_entry_point("query_preprocess")
     # workflow.add_edge("query_preprocess_lambda","intention_detection_lambda")
-    workflow.add_edge("intention_detection_lambda","agent_lambda")
-    workflow.add_edge("tool_execute_lambda","agent_lambda")
-    workflow.add_edge("agent_lambda",'parse_tool_calling')
-    workflow.add_edge("rag_daily_reception_retriever","rag_daily_reception_llm")
-    workflow.add_edge('rag_goods_exchange_retriever',"rag_goods_exchange_llm")
-    workflow.add_edge('rag_product_aftersales_retriever',"rag_product_aftersales_llm")
-    workflow.add_edge('rag_customer_complain_retriever',"rag_customer_complain_llm")
-    workflow.add_edge('rag_promotion_retriever',"rag_promotion_llm")
+    workflow.add_edge("intention_detection","agent")
+    workflow.add_edge("tool_execute","agent")
+    # workflow.add_edge("agent",'parse_tool_calling')
+    # workflow.add_edge("rag_daily_reception_retriever","rag_daily_reception_llm")
+    # workflow.add_edge('rag_goods_exchange_retriever',"rag_goods_exchange_llm")
+    # workflow.add_edge('rag_product_aftersales_retriever',"rag_product_aftersales_llm")
+    # workflow.add_edge('rag_customer_complain_retriever',"rag_customer_complain_llm")
+    # workflow.add_edge('rag_promotion_retriever',"rag_promotion_llm")
     workflow.add_edge('final_rag_retriever',"final_rag_llm")
     
     # end
-    workflow.add_edge("transfer_reply",END)
-    workflow.add_edge("give_rhetorical_question",END)
+    # workflow.add_edge("transfer_reply",END)
+    # workflow.add_edge("give_rhetorical_question",END)
     # workflow.add_edge("give_response_wo_tool",END)
-    workflow.add_edge("rag_daily_reception_llm",END)
-    workflow.add_edge("rag_goods_exchange_llm",END)
-    workflow.add_edge("rag_product_aftersales_llm",END)
-    workflow.add_edge("rag_customer_complain_llm",END)
-    workflow.add_edge('rule_url_reply',END)
+    # workflow.add_edge("rag_daily_reception_llm",END)
+    # workflow.add_edge("rag_goods_exchange_llm",END)
+    # workflow.add_edge("rag_product_aftersales_llm",END)
+    # workflow.add_edge("rag_customer_complain_llm",END)
+    # workflow.add_edge('rule_url_reply',END)
     workflow.add_edge('rule_number_reply',END)
-    workflow.add_edge("rag_promotion_llm",END)
-    workflow.add_edge("give_final_response",END)
+    # workflow.add_edge("rag_promotion_llm",END)
+    # workflow.add_edge("give_final_response",END)
     workflow.add_edge("final_rag_llm",END)
 
     # temporal add edges for ending logic
@@ -831,16 +799,16 @@ def build_graph():
         agent_route,
         {
             "invalid tool calling": "agent_lambda",
-            "rhetorical question": "give_rhetorical_question",
-            "transfer": "transfer_reply",
-            "goods exchange": "rag_goods_exchange_retriever",
-            "daily reception": "rag_daily_reception_retriever",
-            "product aftersales": "rag_product_aftersales_retriever",
-            "customer complain": "rag_customer_complain_retriever",
-            "promotion": "rag_promotion_retriever",
-            "give final response": "give_final_response",
+            # "rhetorical question": "give_rhetorical_question",
+            # "transfer": "transfer_reply",
+            # "goods exchange": "rag_goods_exchange_retriever",
+            # "daily reception": "rag_daily_reception_retriever",
+            # "product aftersales": "rag_product_aftersales_retriever",
+            # "customer complain": "rag_customer_complain_retriever",
+            # "promotion": "rag_promotion_retriever",
+            # "give final response": "give_final_response",
             "final rag": "final_rag_retriever",
-            "continue": "tool_execute_lambda",
+            "continue": "tool_execute",
             
         }
     )
@@ -873,15 +841,21 @@ def retail_entry(event_body):
     :param event_body: The event body for lambda function.
     return: answer(str)
     """
-    global app 
+    global app,app_agent
     if app is None:
-        app = build_graph()
+        app = build_graph(ChatbotState)
+    
+    if app_agent is None:
+        app_agent = build_agent_graph(ChatbotState)
 
     # debuging
     # TODO only write when run local
     if is_running_local():
         with open('retail_entry_workflow.png','wb') as f:
             f.write(app.get_graph().draw_mermaid_png())
+        
+        with open('retail_entry_agent_workflow.png','wb') as f:
+            f.write(app_agent.get_graph().draw_mermaid_png())
     ################################################################################
     # prepare inputs and invoke graph
     event_body['chatbot_config'] = parse_retail_entry_config(event_body['chatbot_config'])
